% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-PA.R
\name{AFForList}
\alias{AFForList}
\title{Computes AF given baseline and scenario}
\usage{
AFForList(scenario, baseline)
}
\value{
A list of AFs stratified by age and sex
}
\description{
Computes AF given baseline and scenario RRs relative to baseline.
}

