% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{AFForList2}
\alias{AFForList2}
\title{Computes AF given baseline and scenario}
\usage{
AFForList2(scenario, baseline)
}
\arguments{
\item{scenario}{RR compared with no exposure}

\item{baseline}{RR compared with no exposure}
}
\value{
A list of AFs stratified by age and sex
}
\description{
Computes AF given baseline and scenario RRs relative to baseline.
}

