% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ParameterSet.R
\docType{methods}
\name{computeAF}
\alias{computeAF}
\alias{computeAF,ParameterSet,ParameterSet-method}
\title{Retrieves the quantiles of the mixture distibution}
\usage{
computeAF(baseline, scenario)

\S4method{computeAF}{ParameterSet,ParameterSet}(baseline, scenario)
}
\arguments{
\item{baseline}{ps}

\item{scenario}{ps}
}
\value{
Foo
}
\description{
Retrieves the quantiles of the mixture distibution
}
\note{
For a list of parameter names use \code{getParameterNames}.
}

