% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{computeMultiplier}
\alias{computeMultiplier}
\title{Computes road injury multiplier}
\usage{
computeMultiplier(base, scenario, safetyInNumbers)
}
\arguments{
\item{base}{a list containing two data frames (one for person-miles
and one for vehicle miles) indicating the distance travelled in
the baseline}

\item{scenario}{Same as above but for the scenario.}
}
\value{
A list of multipliers.
}
\description{
This multiplier is used to estimate the road injury count in the
scenario.  It is a function of distance by road type (person-miles
and vehicle-miles) and safey in number parameters (one for
person-miles and one for vehicle miles.  Currently both values are
fixed at 0.5 for both person and vehicle.
}

