% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{computeQuintiles}
\alias{computeQuintiles}
\title{Compute Quintiles of the Lognormal Distribution}
\usage{
computeQuintiles(mean, sd)
}
\arguments{
\item{mean}{A numerical matrix of means for active transport time}

\item{sd}{A numerical matrix of standard deviations for active
transport time}
}
\value{
A vector of quintiles
}
\description{
Compute quintiles of the lognormal distribution given matrices for
mean and standard deviation of active transport time
}
\note{
This function needs to be cleaned up so it is more user
    friendly

Quintiles are defined as 10, 30, 50, 70, 90 percentiles
}

