% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{computeWalkingMETs}
\alias{computeWalkingMETs}
\title{Compute METs given walking speed}
\usage{
computeWalkingMETs(v)
}
\value{
An estimate for MET expenditure
}
\description{
Compute METs given walking speed
}

