% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{convertAge}
\alias{convertAge}
\title{This function coverts the 2001 CDC Age variable to years.}
\usage{
convertAge(value, unit)
}
\arguments{
\item{value}{The value for age, i.e., characters 2-4 of the code
(sloppy with class!)}

\item{unit}{The units for age, i.e., chatater 1 of the code (sloppy
with class!)}
}
\value{
A numeric vector of age in years - this is a test
}
\description{
This function coverts the 2001 CDC Age variable to years.
}
\note{
This function is likely only applicable to the 2001 data
    file.
}

