% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-auxillary.R
\name{convertToAgeClass}
\alias{convertToAgeClass}
\title{Converts age to ITHIM age class}
\usage{
convertToAgeClass(age)
}
\arguments{
\item{age}{A vector of ages.  This vector will be coerced to numeric.}
}
\value{
A character vector of ITHIM age categories
}
\description{
Converts age to ITHIM age class
}

