% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRA.R
\name{CRA.DF}
\alias{CRA.DF}
\title{Perform comparative risk assesment}
\usage{
CRA.DF(leisure = "medium", n = 10000)
}
\arguments{
\item{n}{Number of quantiles (parametric)}

\item{leisue}{A character string, either "low", "medium" or "high"
indicating the amonut of leisure activity on the population}
}
\value{
A data frame with population attributable fractions
}
\description{
Foo
}
