% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRA.R
\name{CRA}
\alias{CRA}
\title{Perform comparative risk assesment}
\usage{
CRA(activityFile.baseline = system.file("activity.baseline.csv", package =
  "ITHIM"), activityFile.scenario = system.file("activity.scenario.csv",
  package = "ITHIM"))
}
\arguments{
\item{activityFile.baseline}{Name of activity file for baseline}

\item{activityFile.scenario}{Name of activity file for scenario}
}
\value{
A data frame with population attributable fractions
}
\description{
Foo
}
