% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createAirPollutionRRs}
\alias{createAirPollutionRRs}
\title{Set Risk Ratios for Air Pollution}
\usage{
createAirPollutionRRs()
}
\value{
A numerical vector of risk ratios given air pollution exposure
}
\description{
Set risk ratios for a list of diseases given air pollution exposure.  These
values are used to compute change in disease burden due to air pollution.
}
\note{
Hypertensive HD is done using a combination of RR by METs and RR by air
    pollution.
}

