% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createATandAPRRs}
\alias{createATandAPRRs}
\title{Set Risk Ratios for Active Transport and Air Pollution}
\usage{
createATandAPRRs()
}
\value{
A numerical vector of risk ratios given air pollution exposure and active transport exposure
}
\description{
Set risk ratios for a list of diseases given air pollution exposure and active transport exposure.  These
values are used to compute change in disease burden..
}

