% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-createITHIM.R
\docType{methods}
\name{createITHIM}
\alias{createITHIM}
\alias{createITHIM,character,character,character,character,character-method}
\alias{createITHIM,character,character,missing,character,character-method}
\alias{createITHIM,character,character,missing,missing,missing-method}
\alias{createITHIM,character,missing,character,character,character-method}
\alias{createITHIM,character,missing,missing,character,character-method}
\alias{createITHIM,missing,missing,missing,missing,missing-method}
\title{Create an ITHIM object}
\usage{
createITHIM(activeTransportFile, GBDFile, roadInjuriesFile, distRoadTypeFile,
  safetyInNumbersFile)


  \S4method{createITHIM}{character,character,character,character,character}(activeTransportFile,
  GBDFile, roadInjuriesFile, distRoadTypeFile, safetyInNumbersFile)


  \S4method{createITHIM}{character,character,missing,character,character}(activeTransportFile,
  GBDFile, roadInjuriesFile, distRoadTypeFile, safetyInNumbersFile)


  \S4method{createITHIM}{character,character,missing,missing,missing}(activeTransportFile,
  GBDFile, roadInjuriesFile, distRoadTypeFile, safetyInNumbersFile)


  \S4method{createITHIM}{character,missing,missing,character,character}(activeTransportFile,
  GBDFile, roadInjuriesFile, distRoadTypeFile, safetyInNumbersFile)


  \S4method{createITHIM}{character,missing,character,character,character}(activeTransportFile,
  GBDFile, roadInjuriesFile, distRoadTypeFile, safetyInNumbersFile)


  \S4method{createITHIM}{missing,missing,missing,missing,missing}(activeTransportFile,
  GBDFile, roadInjuriesFile, distRoadTypeFile, safetyInNumbersFile)
}
\arguments{
\item{activeTransportFile}{A character string indicating the name
of the file containing mean walk and cycle times.  Default
value uses data from the inst directory.}

\item{GBDFile}{A character string indicating the name of the file
containing the disease burden data.  Default value uses data
from the inst directory.}

\item{roadInjuriesFile}{A character string indicating the name of
the file containing the road injury data (still under
development).  Default value uses data from the inst directory.}

\item{distRoadTypeFile}{A character string indicating the name of
the file containing distances (person and vehicle) by road type
(minor, major, motorOnly)}

\item{safetyInNumbersFile}{A character string indicating the name
of the file conatining safety in number parameters.}
}
\value{
An object of class ITHIM
}
\description{
Returns an ITHIM object.
}
\note{
The path is relative to current working directory.

GBD stands for Global Burden of Disease.  We use this simply
    because this was our original data source.  It does not need to
    be estimates from GBD.  Any disease burden estimate is fine.

The disease burden file must contain values for any of the
    following diseases; "BreastCancer", "ColonCancer", "CVD",
    "Dementia", "Diabetes", "Depression" and "RTIs", and must be
    labeled accordingly.  RTI stands for road traffic injury and is
    needed only if the road injury component is being used.

The format for the disease burden will be normalized and
    explained in greater detail soon.  See
    \url{https://github.com/syounkin/ITHIM/blob/devel/inst/gbd.csv}
    for an example.

If run with no arguments this function will return the
    default ITHIM object
}
\examples{

activeTransportFile <- system.file("activeTransportTime.csv", package = "ITHIM")
GBDFile <- system.file("gbd.csv", package = "ITHIM")

ITHIM <- createITHIM(activeTransportFile = activeTransportFile, GBDFile = GBDFile)

}

