% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createITHIM}
\alias{createITHIM}
\title{Create ITHIM object}
\usage{
createITHIM(vision = "baseline", region = "national")
}
\arguments{
\item{vision}{either "baseline" or "scenario"}

\item{region}{either "national" or "SFBayArea"}
}
\value{
A list of parameters, means and quintiles.
}
\description{
Create ITHIM object
}

