% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createITHIMFunction}
\alias{createITHIMFunction}
\title{Create an ITHIM object}
\usage{
createITHIMFunction()
}
\value{
A list of parameters, means and quintiles.
}
\description{
An ITHIM object is a list which contains three elements;
parameters, means and quintiles.  The parameters are listed in
\code{\link{createParameterList}}.  The elements means and
quintiles mimic the computation presented in the original EXCEL
workbook.  Use \code{\link{updateITHIM}} to change values of the
parameters.
}
\seealso{
\code{\link{updateITHIM}},
    \code{\link{createParameterList}},
    \code{\link{computeMeanMatrices}}, \code{\link{getQuintiles}}
}

