% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ITHIM.R
\name{createParameterList}
\alias{createParameterList}
\title{Define ITHIM model parameters}
\usage{
createParameterList(roadInjuriesFile = system.file("roadInjuries.csv", package
  = "ITHIM"), activeTransportTimeFile = system.file("activeTransportTime.csv",
  package = "ITHIM"), GBDFile = system.file("gbd.csv", package = "ITHIM"))
}
\value{
A list with parameters and estimates
}
\description{
This function is used to generate a complete list of default
parameters
}
\seealso{
\code{\link{readGBD}}
}

