% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createRegionalParameters}
\alias{createRegionalParameters}
\title{Creates a list of Regional Parameters}
\usage{
createRegionalParameters(region, nAgeClass)
}
\arguments{
\item{region}{See \code{\link{createParameterList}}}
}
\value{
A list with four ITHIM parameters

\item{F}{A numerical matrix for the population density, stratified by age class and sex}
\item{Rwt}{A numerical matrix for the walking time, relative to referent group}
\item{Rws}{A numerical matrix for the walking speed, relative to referent group}
\item{Rct}{A numerical matrix for the cycling time, relative to referent group}
}
\description{
Creates a list of Regional Parameters
}
\note{
The referent group is women age class 3.
}

