% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ITHIM.R
\docType{methods}
\name{deltaBurden}
\alias{deltaBurden}
\alias{deltaBurden,ITHIM,ITHIM,character,character,character-method}
\alias{deltaBurden,ITHIM,ITHIM,character,character,missing-method}
\alias{deltaBurden,ITHIM,ITHIM,character,missing,character-method}
\alias{deltaBurden,ITHIM,ITHIM,character,missing,missing-method}
\alias{deltaBurden,ITHIM,ITHIM,missing,character,character-method}
\alias{deltaBurden,ITHIM,ITHIM,missing,character,missing-method}
\alias{deltaBurden,ITHIM,ITHIM,missing,missing,character-method}
\alias{deltaBurden,ITHIM,ITHIM,missing,missing,missing-method}
\title{Gets Change in Disease Burden due to Physical Activity}
\usage{
deltaBurden(baseline, scenario, bur, dis, type)

\S4method{deltaBurden}{ITHIM,ITHIM,character,character,character}(baseline,
  scenario, bur, dis, type)

\S4method{deltaBurden}{ITHIM,ITHIM,character,character,missing}(baseline,
  scenario, bur, dis, type)

\S4method{deltaBurden}{ITHIM,ITHIM,character,missing,character}(baseline,
  scenario, bur, dis, type)

\S4method{deltaBurden}{ITHIM,ITHIM,character,missing,missing}(baseline,
  scenario, bur, dis, type)

\S4method{deltaBurden}{ITHIM,ITHIM,missing,character,character}(baseline,
  scenario, bur, dis, type)

\S4method{deltaBurden}{ITHIM,ITHIM,missing,character,missing}(baseline,
  scenario, bur, dis, type)

\S4method{deltaBurden}{ITHIM,ITHIM,missing,missing,character}(baseline,
  scenario, bur, dis, type)

\S4method{deltaBurden}{ITHIM,ITHIM,missing,missing,missing}(baseline, scenario,
  bur, dis, type)
}
\arguments{
\item{baseline}{ITHIM object}

\item{scenario}{ITHIM object}

\item{bur}{A character string indicating the type of disease burden
measure.  Available values are "deaths", "yll", "yld" and
"daly".  The default value is "daly".}

\item{dis}{A character string indicating which disease to consider.
Possible values are "BreastCancer", "ColonCancer",
"Depression", "Dementia", "Diabetes" and "CVD".  The default
value is "all" which returns the sum of disease burden across
these diseases.}

\item{type}{A character string.  Either "percent" or "absolute",
default is "absolute"}
}
\value{
A burden value
}
\description{
Gets Change in Disease Burden due to Physical Activity
}
\note{
The parameters dis and bur should be harmonized across
    \code{getBurden} and \code{deltaBurden}.
}

