% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-PA.R
\name{deltaBurden}
\alias{deltaBurden}
\title{Performs the comparitive risk assesment for the physical activity
component}
\usage{
deltaBurden(ITHIM.baseline, ITHIM.scenario, bur = "daly.delta",
  dis = "total")
}
\arguments{
\item{ITHIM.baseline}{An ITHIM object}

\item{ITHIM.scenario}{An ITHIM object}

\item{bur}{A chatacter string indicating the type of burden.
Acceptable values are daly.delta, yll.delta, yld.delta,
dproj.delta (projected deaths).}

\item{dis}{A character string indicating which disease is of interest.  Acceptable values are BreastCancer, ColonCancer, CVD, Dementia, Depression, Diabetes, Stroke, HHD or total.  The default value is tot}
}
\value{
A numerical value for the chnage in disease burden between
    baseline and scenario.  Physical activity component only.
}
\description{
This function performs the CRA and returns the change in disease burden.  The user may specify among four burden types and disease (or total).
}

