% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ITHIM.R
\docType{methods}
\name{getAF}
\alias{getAF}
\alias{getAF,ITHIM,ITHIM-method}
\title{Foo}
\usage{
getAF(baseline, scenario)

\S4method{getAF}{ITHIM,ITHIM}(baseline, scenario)
}
\arguments{
\item{baseline}{ITHIM object}

\item{scenario}{ITHIM object}
}
\value{
A list of attributable fractions
}
\description{
Foobar
}

