% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ITHIM.R
\docType{methods}
\name{getCycleTime}
\alias{getCycleTime}
\alias{getCycleTime,ITHIM,integer-method}
\alias{getCycleTime}
\alias{getCycleTime,ITHIM,numeric-method}
\alias{getCycleTime}
\alias{getCycleTime,ITHIM,missing-method}
\alias{getCycleTime}
\title{Returns the mean cycle time matrix}
\usage{
getCycleTime(x, form)

\S4method{getCycleTime}{ITHIM,integer}(x, form)

\S4method{getCycleTime}{ITHIM,numeric}(x, form)

\S4method{getCycleTime}{ITHIM,missing}(x, form)
}
\arguments{
\item{x}{An ITHIM object}

\item{form}{an integer}
}
\value{
A numerical matrix of mean cycle time
}
\description{
Returns the mean cycle time matrix
}
