% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getCycleTime}
\alias{getCycleTime}
\title{Returns the mean cycle time matrix}
\usage{
getCycleTime(x, form)
}
\arguments{
\item{x}{An ITHIM object}

\item{form}{an integer}
}
\value{
A numerical matrix of mean cycle time
}
\description{
Returns the mean cycle time matrix
}

