% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ITHIM.R, R/methods-ParameterSet.R
\docType{methods}
\name{getDistRoadType}
\alias{getDistRoadType}
\alias{getDistRoadType,ITHIM-method}
\alias{getDistRoadType,ParameterSet-method}
\title{Retrieves the distance by road type}
\usage{
getDistRoadType(x)

\S4method{getDistRoadType}{ITHIM}(x)

\S4method{getDistRoadType}{ParameterSet}(x)
}
\arguments{
\item{x}{An ITHIM object or a ParameterSet object}
}
\value{
A list of matrices...
}
\description{
Retrieve distance by road type contained in ITHIM object.
}
\note{
The road injury component of the ITHIM package is still under
    development.  It is not recommended to use this method right
    now.
}

