% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ITHIM.R, R/methods-ParameterSet.R
\docType{methods}
\name{getF}
\alias{getF}
\alias{getF,ITHIM-method}
\alias{getF,ParameterSet-method}
\title{Get F}
\usage{
getF(x)

\S4method{getF}{ITHIM}(x)

\S4method{getF}{ParameterSet}(x)
}
\arguments{
\item{x}{ITHIM or ParameterSetobject}
}
\value{
foo
}
\description{
Get F
}

