% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getGBD}
\alias{getGBD}
\title{Retrieves the Global Burden of Disease (GBD) data}
\usage{
getGBD(x, format = "data.frame")
}
\arguments{
\item{x}{An ITHIM object}

\item{format}{A character string indicating what form to return the
data.  Possible values are "data.frame" and "list".}
}
\value{
Either a list or a data frame containg GBD data.
}
\description{
Retrieves the Global Burden of Disease (GBD) data from an ITHIM object
}

