% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getLogNormal}
\alias{getLogNormal}
\title{Retrieve a density vector for a logNormal distribution}
\usage{
getLogNormal(mu, sd)
}
\arguments{
\item{mu}{The mean (on log scale or not?  Figure this out.}

\item{sd}{The standard deviation (on log scale or not?  Figure this out.}
}
\value{
A vector of length 2000 with density values over the
    interval 0 to 2000
}
\description{
Retrieve a density vector for a logNormal distribution
}

