% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ITHIM.R, R/methods-ParameterSet.R
\docType{methods}
\name{getMeans}
\alias{getMeans}
\alias{getMeans,ITHIM-method}
\alias{getMeans,ParameterSet-method}
\title{Retrieve mean values}
\usage{
getMeans(x)

\S4method{getMeans}{ITHIM}(x)

\S4method{getMeans}{ParameterSet}(x)
}
\arguments{
\item{x}{An ITHIM object}
}
\value{
A numeric vector of means.
}
\description{
Retrieves walking, cycling and nonTravel activity means from an
ITHIM object.
}

