% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getMeans}
\alias{getMeans}
\title{Retrieve mean values}
\usage{
getMeans(x)
}
\arguments{
\item{x}{An ITHIM object}
}
\value{
A numeric vector of means.
}
\description{
Retrieves walking, cycling and nonTravel activity means from an
ITHIM object.
}

