% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-auxillary-exported.R
\name{getMethodsITHIM}
\alias{getMethodsITHIM}
\title{Retrieves the names of methods for ITHIM Objects}
\usage{
getMethodsITHIM()
}
\value{
A character vector of method names
}
\description{
Retrieves the names of methods for ITHIM Objects
}
