% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getMETQuintiles}
\alias{getMETQuintiles}
\title{Computes MET quintiles}
\usage{
getMETQuintiles(mu, p, cv, size = 1000)
}
\value{
A vector of quintiles.
}
\description{
Uses a mixture distribution to estimate non-travel METs
}

