% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getNonTravelDistribution}
\alias{getNonTravelDistribution}
\title{Computes simulated distribution of nonTravel METs}
\usage{
getNonTravelDistribution(mu, cv, size = 10000)
}
\value{
A random sample from the distribution.
}
\description{
Computes simulated distribution of nonTravel METs
}

