% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ITHIM.R
\docType{methods}
\name{getNonTravelMETs}
\alias{getNonTravelMETs}
\alias{getNonTravelMETs,ITHIM,integer-method}
\alias{getNonTravelMETs}
\alias{getNonTravelMETs,ITHIM,numeric-method}
\alias{getNonTravelMETs}
\alias{getNonTravelMETs,ITHIM,missing-method}
\alias{getNonTravelMETs}
\title{Returns the non-travel METs matrix}
\usage{
getNonTravelMETs(x, form)

\S4method{getNonTravelMETs}{ITHIM,integer}(x, form)

\S4method{getNonTravelMETs}{ITHIM,numeric}(x, form)

\S4method{getNonTravelMETs}{ITHIM,missing}(x, form)
}
\arguments{
\item{x}{An ITHIM object}

\item{form}{an integer}
}
\value{
A numerical matrix of mean non-travel activity
}
\description{
Returns the non-travel METs matrix
}
