% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getNonTravelMETs}
\alias{getNonTravelMETs}
\title{Returns the non-travel METs matrix}
\usage{
getNonTravelMETs(x, form)
}
\arguments{
\item{x}{An ITHIM object}

\item{form}{an integer}
}
\value{
A numerical matrix of mean non-travel activity
}
\description{
Returns the non-travel METs matrix
}

