% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ITHIM.R, R/methods-ParameterSet.R
\docType{methods}
\name{getParameterNames}
\alias{getParameterNames}
\alias{getParameterNames,ITHIM-method}
\alias{getParameterNames,ParameterSet-method}
\title{Returns the names of the parameters in the ITHIM object}
\usage{
getParameterNames(x)

\S4method{getParameterNames}{ITHIM}(x)

\S4method{getParameterNames}{ParameterSet}(x)
}
\arguments{
\item{x}{ITHIM object or a ParameterSet object}
}
\value{
A character vector of parameter names
}
\description{
Returns the names of the parameters in the ITHIM object
}

