% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getParameterNames}
\alias{getParameterNames}
\title{Returns the names of the parameters in the ITHIM object}
\usage{
getParameterNames(x)
}
\arguments{
\item{x}{ITHIM object or a ParameterSet object}
}
\value{
A character vector of parameter names
}
\description{
Returns the names of the parameters in the ITHIM object
}

