% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ITHIM.R
\docType{methods}
\name{getParameterSet}
\alias{getParameterSet}
\alias{getParameterSet,ITHIM-method}
\title{Retrieve a ParameterSet object}
\usage{
getParameterSet(x)

\S4method{getParameterSet}{ITHIM}(x)
}
\arguments{
\item{x}{An ITHIM object}
}
\value{
An object of class ParameterSet
}
\description{
Returns the ParameterSet object embedded in an ITHIM object.
}

