% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getParameterSet}
\alias{getParameterSet}
\title{Retrieve a ParameterSet object}
\usage{
getParameterSet(x)
}
\arguments{
\item{x}{An ITHIM object}
}
\value{
An object of class ParameterSet
}
\description{
Returns the ParameterSet object embedded in an ITHIM object.
}

