% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-auxilliary.R
\docType{methods}
\name{getPopulation}
\alias{getPopulation}
\alias{getPopulation,vector,vector-method}
\title{Retrieves population matrix from U.S. Census}
\usage{
getPopulation(state, county)

\S4method{getPopulation}{vector,vector}(state, county)
}
\arguments{
\item{state}{An integer representing the state ID}

\item{county}{An integer representing the county ID}
}
\value{
A data frame of population counts
}
\description{
Retrieves population matrix from U.S. Census
}
\note{
The underlying function, \code{getTractAgeSex}, was written
    primarily by Vargo
}

