% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ParameterSet.R
\docType{methods}
\name{getQuantilesMixture}
\alias{getQuantilesMixture}
\alias{getQuantilesMixture,ParameterSet-method}
\title{Retrieves the quantiles of the mixture distibution}
\usage{
getQuantilesMixture(object)

\S4method{getQuantilesMixture}{ParameterSet}(object)
}
\arguments{
\item{object}{A ParameterSet object}
}
\value{
Foo
}
\description{
Retrieves the quantiles of the mixture distibution
}
\note{
For a list of parameter names use \code{getParameterNames}.
}

