% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getQuintiles}
\alias{getQuintiles}
\title{Compute Quintiles of Active Transport Time}
\usage{
getQuintiles(means)
}
\arguments{
\item{means}{A list of means generated by the \code{\link{computeMeanMatrices}}}
}
\value{
A list of lists containing quintiles of active transport
    time and METs, by sex and age class.

\item{ActiveTransportTime}{Foo}
\item{WalkingTime}{Foo}
\item{CyclingTime}{Foo}
\item{WalkingMET}{Foo}
\item{CyclingMET}{foo}
\item{TotalTravelMET}{foo}
}
\description{
Compute Quintiles of Active Transport Time
}
\seealso{
\code{\link{computeMeanMatrices}}
}

