% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ITHIM.R,
%   R/methods-ParameterSet.R
\docType{methods}
\name{getRoadInjuries}
\alias{getRoadInjuries}
\alias{getRoadInjuries,ITHIM-method}
\alias{getRoadInjuries}
\alias{getRoadInjuries,ParameterSet-method}
\alias{getRoadInjuries}
\title{Gets Road Injury counts from ITHIM object}
\usage{
getRoadInjuries(x)

\S4method{getRoadInjuries}{ITHIM}(x)

\S4method{getRoadInjuries}{ParameterSet}(x)
}
\arguments{
\item{x}{ITHIM object}
}
\value{
A list of matrices of road injury counts.
}
\description{
Gets Road Injury counts from ITHIM object
}
\note{
The matrices have striking modes as rows and victim modes as
    columns.  NOV is considered a striking mode.
}
