% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getRoadInjuries}
\alias{getRoadInjuries}
\title{Gets Road Injury counts from ITHIM object}
\usage{
getRoadInjuries(x)
}
\arguments{
\item{x}{ITHIM object}
}
\value{
A list of matrices of road injury counts.
}
\description{
Gets Road Injury counts from ITHIM object
}
\note{
The matrices have striking modes as rows and victim modes as
    columns.  NOV is considered a striking mode.
}

