% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ITHIM.R,
%   R/methods-ParameterSet.R
\docType{methods}
\name{getSiN}
\alias{getSiN}
\alias{getSiN,ITHIM-method}
\alias{getSiN}
\alias{getSiN,ParameterSet-method}
\alias{getSiN}
\title{Retrieves the Safety in Numbers parameters}
\usage{
getSiN(x)

\S4method{getSiN}{ITHIM}(x)

\S4method{getSiN}{ParameterSet}(x)
}
\arguments{
\item{x}{An ITHIM or ParameterSet object}
}
\value{
An array of safety in numbers parameters
}
\description{
Retrieves the Safety in Numbers parameters
}
