% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getSiN}
\alias{getSiN}
\title{Retrieves the Safety in Numbers parameters}
\usage{
getSiN(x)
}
\arguments{
\item{x}{An ITHIM or ParameterSet object}
}
\value{
An array of safety in numbers parameters
}
\description{
Retrieves the Safety in Numbers parameters
}

