% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsFunctions.R
\name{getTA}
\alias{getTA}
\title{Compute Travel Activity Parameters}
\usage{
getTA(ts, alpha = 1)
}
\arguments{
\item{ts}{A TravelSurvey object}
}
\value{
A data frame with travel activity mean and standard deviation
}
\description{
Foo
}
