% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ParameterSet.R
\docType{methods}
\name{getTravelActivity}
\alias{getTravelActivity}
\alias{getTravelActivity,ParameterSet-method}
\title{Retrieves the mean travel activity}
\usage{
getTravelActivity(object)

\S4method{getTravelActivity}{ParameterSet}(object)
}
\arguments{
\item{object}{A ParameterSet object}
}
\value{
Mean travel activity
}
\description{
Retrieves the mean travel activity
}
\note{
For a list of parameter names use \code{getParameterNames}.
}

