% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getTravelDistribution}
\alias{getTravelDistribution}
\title{Computes simulated distribution of Travel METs}
\usage{
getTravelDistribution(mu, cv, pWalk, vWalk, size = 10000)
}
\value{
A random sample from the distribution.
}
\description{
Computes simulated distribution of Travel METs
}

