% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{getWalkSpeed}
\alias{getWalkSpeed}
\title{Returns the mean walk speed matrix}
\usage{
getWalkSpeed(ITHIM, form = 2)
}
\arguments{
\item{ITHIM}{An ITHIM object}
}
\value{
A numerical matrix of mean walk speed
}
\description{
Returns the mean walk speed matrix
}

