% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ITHIM.R
\docType{methods}
\name{getWalkTime}
\alias{getWalkTime}
\alias{getWalkTime,ITHIM,integer-method}
\alias{getWalkTime,ITHIM,missing-method}
\alias{getWalkTime,ITHIM,numeric-method}
\title{Returns the mean walk time matrix}
\usage{
getWalkTime(x, form)

\S4method{getWalkTime}{ITHIM,integer}(x, form)

\S4method{getWalkTime}{ITHIM,numeric}(x, form)

\S4method{getWalkTime}{ITHIM,missing}(x, form)
}
\arguments{
\item{x}{An ITHIM object}

\item{form}{an integer}
}
\value{
A numerical matrix of mean walk time
}
\description{
Returns the mean walk time matrix
}

