% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getWalkTime}
\alias{getWalkTime}
\title{Returns the mean walk time matrix}
\usage{
getWalkTime(x, form)
}
\arguments{
\item{x}{An ITHIM object}

\item{form}{an integer}
}
\value{
A numerical matrix of mean walk time
}
\description{
Returns the mean walk time matrix
}

