% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{ITHIM-package}
\alias{ITHIM-package}
\title{Integrated Transport and Health Impacts Model (ITHIM)}
\description{
Implementation in R
}
\examples{

ITHIMParameterList <- createParameterList()
meansList <- computeMeanMatrices(ITHIMParameterList)
names(meansList)
meansList$meanActiveTransportTime

}
\author{
Samuel G. Younkin \email{syounkin@wisc.edu}
}
\seealso{
\code{\link{createParameterList}}, \code{\link{computeMeanMatrices}}
}

