% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ITHIMList}
\alias{ITHIMList}
\alias{ITHIMList-class}
\title{The ITHIMList class}
\description{
Foo.
}
\details{
This line and the next ones go into the details.
This line thus appears in the details as well.
}
