% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-PA-nonExported.R
\name{MET2RR}
\alias{MET2RR}
\title{Computes the RR for an exposure of x MET}
\usage{
MET2RR(RR, MET)
}
\value{
A list of matrices of quintiles of RR_x stratified by age
    class and sex
}
\description{
We use the transformation RR_x = RR_1^(x^k), where RR_1 is the
relative risk for one MET.
}
\note{
k is fixed at 0.5 for now
}

