% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ParameterSet}
\alias{ParameterSet}
\alias{ParameterSet-class}
\title{The ITHIM ParameterSet class}
\description{
This is the secondary class in the ITHIM package.
}
\details{
This line and the next ones go into the details.
This line thus appears in the details as well.
}
\section{Slots}{

\describe{
\item{\code{}:}{}
}
}

