% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotBurden}
\alias{plotBurden}
\title{Plots burden}
\usage{
plotBurden(burden, varName = "daly")
}
\arguments{
\item{burden}{A list like as "comparativeRisk$daly.delta"}

\item{varName}{A character string indicating the name to use for
the y axis label}
}
\value{
A ggplot object
}
\description{
Plots burdens
}

