% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{readGBD}
\alias{readGBD}
\title{Read in Global Burden of Disease Data}
\usage{
readGBD(filename)
}
\value{
A list of lists of matrices with dproj, yll, yld and daly
    by age and sex and disease
}
\description{
Read in Global Burden of Disease Data
}

