% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-GBD.R
\name{reformatGBD}
\alias{reformatGBD}
\title{Reformat GBD data frame}
\usage{
reformatGBD(gbd)
}
\value{
A list of lists of matrices with dproj, yll, yld and daly
    by age and sex and disease
}
\description{
Reformat GBD data frame
}

