% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{sumCVD}
\alias{sumCVD}
\title{Returns}
\usage{
sumCVD(ITHIM.baseline, ITHIM.scenario)
}
\arguments{
\item{ITHIM.baseline}{An ITHIM object}

\item{ITHIM.scenario}{An ITHIM object}
}
\value{

}
\description{
Returns
}

