% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ITHIM.R
\docType{methods}
\name{summariseAF}
\alias{summariseAF}
\alias{summariseAF,ITHIM,ITHIM-method}
\title{Foo}
\usage{
summariseAF(baseline, scenario)

\S4method{summariseAF}{ITHIM,ITHIM}(baseline, scenario)
}
\arguments{
\item{baseline}{ITHIM object}

\item{scenario}{ITHIM object}
}
\value{
A vector of summarised attributable fractions
}
\description{
Foobar
}

