% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ITHIM.R
\name{superTabulate}
\alias{superTabulate}
\title{Tabulate the results of Comparative Risk Assessment}
\usage{
superTabulate(ITHIM.baseline, ITHIM.scenario.list)
}
\arguments{
\item{ITHIM.baseline}{Baseline}

\item{ITHIM.scenario.list}{An ITHIM list}
}
\value{
A data frame of change in burden
}
\description{
Given an ITHIM baseline object and a list of ITHIM scenario objects
this function returns a data frame of change in burden (absolute,
not percent)
}

