% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ITHIM.R
\docType{methods}
\name{tabulateDeltaBurden}
\alias{tabulateDeltaBurden}
\alias{tabulateDeltaBurden,ITHIM,ITHIM-method}
\title{Foo}
\usage{
tabulateDeltaBurden(baseline, scenario)

\S4method{tabulateDeltaBurden}{ITHIM,ITHIM}(baseline, scenario)
}
\arguments{
\item{baseline}{foo}

\item{scenario}{foo}
}
\value{
foo
}
\description{
Foo
}

