% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ITHIM.R
\name{tabulateResults}
\alias{tabulateResults}
\title{Update an ITHIM object}
\usage{
tabulateResults(ITHIM.baseline, ITHIM.scenario.list)
}
\arguments{
\item{ITHIM.baseline}{Baseline}

\item{ITHIM.scenario.list}{An ITHIM list}
}
\value{
A data frame of change in burden
}
\description{
Takes an ITHIM object as input and returns a new ITHIM object with
parameters updated accordingly.
}
