% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{tilePlot}
\alias{tilePlot}
\title{Tile plot}
\usage{
tilePlot(x, n)
}
\arguments{
\item{x}{ITHIM object}

\item{n}{A numeric value indicating the length of the walk and
cycle vectors}
}
\value{
A figure created by ggplot2
}
\description{
Creates a tile plot
}

