% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-ITHIM.R,
%   R/methods-ParameterSet.R
\docType{methods}
\name{update}
\alias{update}
\alias{update,ITHIM,missing-method}
\alias{update}
\alias{update,ITHIM,ParameterSet-method}
\alias{update}
\alias{update,ITHIM,list-method}
\alias{update}
\alias{update,ParameterSet,list-method}
\alias{update}
\title{Update an ITHIM object}
\usage{
update(x, parList)

\S4method{update}{ITHIM,missing}(x, parList)

\S4method{update}{ITHIM,ParameterSet}(x, parList)

\S4method{update}{ITHIM,list}(x, parList)

\S4method{update}{ParameterSet,list}(x, parList)
}
\arguments{
\item{x}{ITHIM object or a ParameterSet object}

\item{parList}{A named list of parameters}
}
\value{
An ITHIM object
}
\description{
Takes an ITHIM object as input and returns a new ITHIM object with
parameters updated accordingly.
}
\note{
For a list of parameter names use \code{getParameterNames}.
}
