% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{update}
\alias{update}
\title{Update an ITHIM object}
\usage{
update(x, parList)
}
\arguments{
\item{x}{ITHIM object}

\item{parList}{A named list of parameters}
}
\value{
An ITHIM object
}
\description{
Takes an ITHIM object as input and returns a new ITHIM object with
parameters updated accordingly.
}
\note{
For a list of parameter names use \code{getParameterNames}.
}
\examples{
ITHIM <- createITHIM()
getMeans(ITHIM)
ITHIM.new <- update(ITHIM, list(muwt = 10, muct = 5))
getMeans(ITHIM.new)

}

