% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ITHIM.R
\name{updateITHIM}
\alias{updateITHIM}
\title{Updates an ITHIM object}
\usage{
updateITHIM(ITHIM, parName, parValue)
}
\value{
An updated ITHIM object
}
\description{
Change a parameter and recreate the object.
}

