% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compareModels}
\alias{compareModels}
\title{Computes AF given baseline and scenario models}
\usage{
compareModels(baseline, scenario, GBDFile = "~/ITHIM/R/gbd.csv")
}
\value{
A list of AFs stratified by age and sex
}
\description{
Computes AF given baseline and scenario
}

