% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{computeNonTravelMETs}
\alias{computeNonTravelMETs}
\title{Generate matrices of non-travel related MET quintiles}
\usage{
computeNonTravelMETs(region)
}
\arguments{
\item{region}{character string; either "national" or "SFBayArea"}
}
\value{
A list of two matrices of quintiles of non-transport METs
    per week stratified by age class and sex
}
\description{
Estimate non-travel related MET quintiles.  Currently we use a
fixed matrix coded into the function.
}

