% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRA.R
\name{CRA}
\alias{CRA}
\title{Perform comparative risk assesment}
\usage{
CRA(meanlog.baseline = 10, sdlog.baseline = 1, p0.baseline = 0.5,
  meanlog.scenario = 10, sdlog.scenario = 1, p0.scenario = 0.5,
  n = 10000, B = 1e+05, P, Q, R = function(x) exp(-x),
  type = "parametric")
}
\arguments{
\item{meanlog.baseline}{Mean of exposure in baseline (log-scale, parameteric model)}

\item{sdlog.baseline}{Standard deviation of exposure in baseline (log-scale, parameteric model)}

\item{p0.baseline}{Proportion of inactive travelers in baseline}

\item{meanlog.scenario}{Mean of exposure in scenario (log-scale, parameteric model)}

\item{sdlog.scenario}{Standard deviation of exposure in baseline (log-scale, parameteric model)}

\item{p0.scenario}{Proportion of inactive travelers in scenario}

\item{n}{Number of quantiles}

\item{B}{Sample size (parametric model)}

\item{P}{Quantiles of exposure in baseline (non-parametric model)}

\item{Q}{Quantiles of exposure in scenario (non-parametric model)}

\item{R}{Relative risk function in terms of total exposure (physical activity)}

\item{type}{A character string.  Either "parametric" or "non-parametric"}
}
\value{
The population attributable fraction
}
\description{
Foo
}

