% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{createITHIM}
\alias{createITHIM}
\title{Create an ITHIM object}
\usage{
createITHIM(x)
}
\arguments{
\item{x}{A list of file names or a ParameterSet object}
}
\value{
An object of class ITHIM
}
\description{
Returns an ITHIM object.
}
\note{
If run with no arguments this function will return the
    default ITHIM object
}

