% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createParameterList}
\alias{createParameterList}
\title{Define ITHIM model parameters}
\usage{
createParameterList()
}
\value{
A list with parameters and estimates

\item{Rwt}{A numerical matrix for the walking time, relative to ?value?}
\item{Rct}{A numerical matrix for the cycling time, relative to ?value?}
\item{Rws}{A numerical matrix for the walking speed, relative to ?value?}
\item{muwt}{A numerical value for the mean walking time}
\item{muws}{A numerical value for the mean walking speed}
\item{muct}{A numerical value for the mean cycling time}
\item{cv}{A numerical value for the coefficient of variation for active transport time}
\item{cvNonTravel}{A numerical value for the coefficient of variation for leisure activity}
\item{muNonTravel}{}
\item{muNonTravelMatrix}{}
\item{GBD}{}
\item{meanType}{}
\item{quantiles}{}
}
\description{
This function is used to generate a complete list of default
parameters
}
\note{
There are 11 parameters in the ITHIM active transport component;

1-4. mean walking and cycling times (muwt, muct, Rwt, Rct)

5. standard deviation of active travel time (cv),

6-7. walk speed (muws, Rws)

8. ratio of regional disease-specific mortality to national disease-specific mortality (GBD)

9-10. non-travel related physical activity means by age and sex (muNonTravel, muNonTravelMatrix)

11. standard deviation of leisure activity (cvNonTravel),
}
\seealso{
\code{\link{computeNonTravelMETs}},\code{\link{readGBD}}
}

