% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createParameterList}
\alias{createParameterList}
\title{Define ITHIM model parameters}
\usage{
createParameterList()
}
\arguments{
\item{region}{A character string, either "national" or "SFBayArea"}
}
\value{
A list with parameters and estimates

\item{F}{A numerical matrix for the population density, stratified by age class and sex}
\item{Rwt}{A numerical matrix for the walking time, relative to ?value?}
\item{Rws}{A numerical matrix for the walking speed, relative to ?value?}
\item{Rct}{A numerical matrix for the cycling time, relative to ?value?}
\item{muwt}{A numerical value for the mean walking time}
\item{muws}{A numerical value for the mean walking speed}
\item{muct}{A numerical value for the mean cycling time}
\item{cv}{A numerical value for the coefficient of variation for active transport time}
}
\description{
This function is used to generate a list of parameters describing
both the region and vision of interest.
}
\note{
There are 15 parameters in the ITHIM model;

Active Transport

1. mean travel distance (muwt,muws,Rws,Rwt)

3. ratio of active travel time to reference group (Rwt, Rct)

4. standard deviation of active travel time (cv),

5. distribution of population by age and sex (F)

8. walk speed (muws)

15. Ratio of daily per capita bicycling to walking time (?)

Disease Burden

6. ratio of regional disease-specific mortality to national disease-specific mortality (-)

7. proportion of colon cancers from colo-rectal cancer (-)

Non-travel METs

9. non-travel related physical activity by age and sex (fixed)

Air Pollution

2. mean travel time by mode (-)

10. Personal travel distance by facility type and travel mode (-)

11. Vehicle distance traveled by facility type and travel mode (-)

12. CO2 emitted per distance traveled by vehicle (-)

14. Emissions of primary and secondary sources of PM 2.5 (-)

Road Injuries

13. Serious and fatal injuries between a striking vehicle and victim vehicle by severity and modes (-)
}
\seealso{
\code{\link{computeNonTravelMETs}},\code{\link{readGBD}}
}

