% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createParameterList}
\alias{createParameterList}
\title{Create a list of input parameters}
\usage{
createParameterList(baseline = TRUE)
}
\arguments{
\item{baseline}{A logical value indicating if we are using the baseline scenario}
}
\value{
A list with parameters and estimates

\item{F}{A numerical matrix for the population density, stratified by age class and sex}
\item{Rwt}{A numerical matrix for the walking time, relative to ?value?}
\item{Rws}{A numerical matrix for the walking speed, relative to ?value?}
\item{Rct}{A numerical matrix for the cycling time, relative to ?value?}
\item{Rcs}{A numerical matrix for the cycling speed, relative to ?value?}
\item{muwt}{A numerical value for the mean walking time}
\item{muws}{A numerical value for the mean walking speed}
\item{muct}{A numerical value for the mean cycling time}
\item{mucs}{A numerical value for the mean cycling speed}
\item{cv}{A numerical value for the coefficient of variation for active transport time}
}
\description{
Create a list of input parameters
}
\note{
A note about the default values.
}

