% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{deltaBurden}
\alias{deltaBurden}
\title{Gets Change in Disease Burden due to Physical Activity}
\usage{
deltaBurden(x, bur, dis)
}
\arguments{
\item{x}{ITHIM object}

\item{bur}{?}

\item{dis}{?}
}
\value{
A burden value
}
\description{
Gets Change in Disease Burden due to Physical Activity
}

