% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getBurden}
\alias{getBurden}
\title{Gets Burden}
\usage{
getBurden(x, bur, dis)
}
\arguments{
\item{x}{ITHIM object}

\item{bur}{A character string indicating the type of disease burden measure.  Possible values are}

\item{dis}{?}
}
\value{
A burden value
}
\description{
Gets burden
}

