% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-accessor.R
\name{getCycleTime}
\alias{getCycleTime}
\title{Returns the mean cycle time matrix}
\usage{
getCycleTime(ITHIM, form = 2)
}
\arguments{
\item{ITHIM}{An ITHIM object}
}
\value{
A numerical matrix of mean cycle time
}
\description{
Returns the mean cycle time matrix
}

