% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{getNonTravelMETs}
\alias{getNonTravelMETs}
\title{Returns the non-travel METs matrix}
\usage{
getNonTravelMETs(ITHIM, form = 2)
}
\arguments{
\item{ITHIM}{An ITHIM object}
}
\value{
A numerical matrix of mean non-travel means
}
\description{
Returns the non-travel METs matrix
}

