% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-auxilliary.R
\docType{methods}
\name{getPopulation}
\alias{getPopulation}
\alias{getPopulation,missing,vector,vector-method}
\alias{getPopulation}
\alias{getPopulation,numeric,missing,missing-method}
\alias{getPopulation}
\alias{getPopulation,list,missing,missing-method}
\alias{getPopulation}
\title{Retrieves population matrix from U.S. Census}
\usage{
getPopulation(msa, state, county)

\S4method{getPopulation}{missing,vector,vector}(msa, state, county)

\S4method{getPopulation}{numeric,missing,missing}(msa, state, county)

\S4method{getPopulation}{list,missing,missing}(msa, state, county)
}
\arguments{
\item{msa}{An integer representing the MSA ID (NHTS specific)}

\item{state}{An integer representing the state ID}

\item{county}{An integer representing the county ID}
}
\value{
A data frame of population counts
}
\description{
Retrieves population matrix from U.S. Census
}
\note{
The underlying function, \code{getTractAgeSex}, was written
    primarily by Vargo
}
