% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getTotalDistribution}
\alias{getTotalDistribution}
\title{Compute distribution of Total METs}
\usage{
getTotalDistribution(muTravel, cvTravel, muNonTravel, cvNonTravel, pWalk, vWalk,
  size)
}
\value{
An estimate for total MET distribution
}
\description{
Compute distribution of Total METs
}

