% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{getWalkTime}
\alias{getWalkTime}
\title{Returns the mean walk time matrix}
\usage{
getWalkTime(ITHIM, form = 2)
}
\arguments{
\item{ITHIM}{An ITHIM object}
}
\value{
A numerical matrix of mean walk time
}
\description{
Returns the mean walk time matrix
}

