% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{ITHIM-package}
\alias{ITHIM-package}
\title{Integrated Transport and Health Impacts Model (ITHIM)}
\description{
Implementation in R
}
\examples{

ITHIM.baseline <- createITHIM(vision = "baseline", region = "SFBayArea")
ITHIM.scenario <- createITHIM(vision = "scenario", region = "SFBayArea")
comparativeRisk <- compareModels(ITHIM.baseline, ITHIM.scenario)
names(ITHIM.baseline)
names(ITHIM.baseline$parameters)
ITHIM.baseline$quintiles$TotalMET
names(comparativeRisk)
comparativeRisk$AF$BreastCancer
}
\author{
Samuel G. Younkin \email{syounkin@wisc.edu}
}
\seealso{
\code{\link{createITHIM}}
}

