% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ITHIM}
\alias{ITHIM}
\alias{ITHIM-class}
\title{The ITHIM class}
\description{
Foo.
}
\details{
This line and the next ones go into the details.
This line thus appears in the details as well.
}

