% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ParameterSet}
\alias{ParameterSet}
\alias{ParameterSet-class}
\title{The ITHIM ParameterSet class}
\description{
This class defines the set of parameters needed for an
ITHIM analysis.  See ITHIM class for more information.
}

