% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{ratioForList}
\alias{ratioForList}
\title{Computes a ratio of elements in two lists}
\usage{
ratioForList(baseline, scenario)
}
\value{
A list of ratios
}
\description{
Computes a ratio of elements in two lists.
}

