% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-PA.R
\name{readActiveTransportTime}
\alias{readActiveTransportTime}
\title{Read Active Transport File}
\usage{
readActiveTransportTime(filename)
}
\arguments{
\item{filename}{A character string with the name of the active transport csv file.}
}
\value{
A list of...
}
\description{
This function reads in a csv file with mean values for walking and
cycling times, stratified by age and sex.  The format of this file
is described at \url{https://ithim.ghi.wisc.edu/}.
}

