% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readGBD}
\alias{readGBD}
\title{Read in Global Burden of Disease Data}
\usage{
readGBD(file = "~/gbd.csv")
}
\value{
A list of lists of matrices with dproj, yll, yld and daly
    by age and sex and disease
}
\description{
Read in Global Burden of Disease Data
}

