% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions2.R
\name{reshapeRR}
\alias{reshapeRR}
\title{Transforms the RR object}
\usage{
reshapeRR(RR, nQuantiles = 5)
}
\value{
A list of two matrices of RRs stratified by age class and
    sex
}
\description{
Transforms the RR object into something more convenient.
}

